server/requirements.txt
etcd3==0.12.0
requests
grpcio==1.32.0